Ceasing development of Entity Unification
When Entites in the world are represented by the same class than
the one defining them, we duplicate data for each Entity that's
unlikely to ever change. In contrast, the similarities between the
old classes is absolutely irrelevant since the produced overhead is
minimal compared to the data duplication after unification.

Also, my hands are typing words.